backend_core/app/deps.py
import os
import hashlib
import warnings
from datetime import datetime, timedelta
from typing import Generator
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from passlib.context import CryptContext
from sqlalchemy.orm import Session

from . import models
from .db import SessionLocal


def get_secret_key() -> str:
    """
    Get secret key from environment with proper validation.
    NEVER use default secret keys in production!
    """
    secret = os.getenv("KUMAR_SECRET_KEY") or os.getenv("KUMAR_SECRET")
    
    if not secret:
        env = os.getenv("ENVIRONMENT", "development")
        if env == "production":
            raise RuntimeError(
                "CRITICAL: Secret key must be set in production! "
                "Set KUMAR_SECRET_KEY environment variable."
            )
        # Development mode - warn but continue
        warnings.warn(
            "⚠️  No secret key set! Using development key. "
            "Set KUMAR_SECRET_KEY for production.",
            RuntimeWarning
        )
        # Use a deterministic key for development (NOT FOR PRODUCTION)
        secret = hashlib.sha256(b"dev-mode-only").hexdigest()
    
    return secret


SECRET_KEY = get_secret_key()
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("TOKEN_EXPIRE_MINUTES", "60"))

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/login")


def get_db() -> Generator:
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()


def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)


def get_password_hash(password):
    return pwd_context.hash(password)


def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt


def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception
    user = db.query(models.User).filter(models.User.username == username).first()
    if user is None:
        raise credentials_exception
    return user


def require_role(*allowed_roles):
    def role_checker(current_user: models.User = Depends(get_current_user)):
        if current_user.role not in allowed_roles:
            raise HTTPException(status_code=403, detail="Insufficient permissions")
        return current_user

    return role_checker
