requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pandas>=2.0.0
openpyxl>=3.0.10
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
pytest>=7.0.0

# Optional (development)
python-dotenv>=1.0.0
sqlmodel>=0.0.8
